JAMES-3194 Inject DTOConverter for every product
Note that is requires splitting guice injections for it
