+++ server/container/guice/guice-common/src/main/java/org/apache/james/modules/server/TaskManagerModule.java
@@ -29,6 +29,8 @@ public class TaskManagerModule extends AbstractModule {
     @Override
     protected void configure() {
         install(new HostnameModule());
+        install(new TaskSerializationModule());
+
         bind(MemoryTaskManager.class).in(Scopes.SINGLETON);
         bind(TaskManager.class).to(MemoryTaskManager.class);
     }
